requirements.txt
Django
djangorestframework
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.2.0
django-storages==1.14.2
psycopg2-binary
reportlab
gunicorn
dotenv
boto3
botocore
drf-yasg